Rename API folder to more generic lib and update imports
