+++ app/_layout.tsx
@@ -9,7 +9,7 @@ import * as SplashScreen from 'expo-splash-screen'
 import { useEffect } from 'react'
 import 'react-native-reanimated'
 import { ApolloProvider } from '@apollo/client'
-import { client as apolloClient } from '@/API/ContentfulAPI'
+import { client as apolloClient } from '@/lib/ContentfulAPI'
 import Colors from '@/constants/Colors'
 import { useColorScheme } from '@/components/useColorScheme'
 import FontAwesome from '@expo/vector-icons/FontAwesome'
